aarch64: fix android build

+++ public/XUnzip.cpp
@@ -4244,12 +4244,12 @@ ZRESULT TUnzip::Unzip(int index,void *dst,unsigned int len,DWORD flags)
 #ifdef _WIN32
 		SetFileTime(h,&ze.ctime,&ze.atime,&ze.mtime);
 #elif defined( ANDROID )
-        struct timespec ts[2];
+	struct timespec ts[2];
         ts[0].tv_sec = ze.atime;
         ts[0].tv_nsec = 0;
         ts[1].tv_sec = ze.mtime;
         ts[1].tv_nsec = 0;
-        utimensat((int)h, NULL, ts, 0);
+        utimensat((intptr_t)h, NULL, ts, 0);
 #else
 		struct timeval tv[2];
 		tv[0].tv_sec = ze.atime;
